Update feedplugins.py
This is more like the intended behavior: modelstring is a string-like tag by which a user can identify a beam model, e.g. "Hamaker-default" or "VSH-v3"; that is, a modeltype + "-" + version (since mdlversep is "-" here), but if there is no mdlversep in the string the default version will have to be looked up and so it is initially set to None.